Add files via upload
Be sure to download the master configuration file from your FPGA board manufacturer. Configure 16 LEDs into two rows of 8 with current limiting resistors and connect them to the PMOD headers of your choice. In the Master file, configure the PMOD headers using "right" and "left" in "SongPlayer" (the top file). "clock" needs a 100MHz clock. EN is to be configured to a button. Configure your mono speaker to "audioOut." If you are using the Nexys7, you will need to configure "aud_sd". "R", "G", "B" can be configured to an onboard RGB LED or through 3 PMOD header pins.